fix sf1000x id data type

+++ src/main/java/com/hybench/pojo/Transfer.java
@@ -14,7 +14,7 @@ import java.util.StringJoiner;
 
 
 public record Transfer
-        (int id,
+        (long id,
          int sourceID,
          int targetID,
          double amount,
